:truck: move oob test to own module
